Update and rename app/src/main/java/ru/fefu/activitytracker/WelcomeScreen.kt to app/src/main/java/ru/fefu/activitytracker/main/WelcomeScreen.kt
