--- app/src/main/java/ru/fefu/activitytracker/WelcomeScreen.kt
@@ -1,71 +1,0 @@
-package ru.fefu.activitytracker
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.view.View
-import android.widget.Button
-import android.widget.TextView
-
-class WelcomeScreen : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_welcome_screen)
-        val btnRegistration = findViewById<Button>(R.id.registrationButton)
-        btnRegistration.setOnClickListener {
-            startActivity(Intent(this, Registration::class.java))
-        }
-        val clickText = findViewById<TextView>(R.id.textQuestion)
-        clickText.setOnClickListener {
-            startActivity(Intent(this, Login::class.java))
-        }
-    }
-}
- 46  app/src/main/java/ru/fefu/activitytracker/fragments/ActivityFragment.kt 
-@@ -0,0 +1,46 @@
-package ru.fefu.activitytracker.fragments
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Adapter
-import android.widget.TableLayout
-import androidx.viewpager.widget.ViewPager
-import androidx.viewpager2.adapter.FragmentStateAdapter
-import androidx.viewpager2.widget.ViewPager2
-
-import com.google.android.material.tabs.TabLayout
-import com.google.android.material.tabs.TabLayoutMediator
-
-import ru.fefu.activitytracker.R
-
-class ActivityFragment : Fragment() {
-    private lateinit var adapter: ViewPagerAdapter
-    private lateinit var viewPager: ViewPager2
-    private lateinit var tabLayout: TableLayout
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        val myFragment = inflater.inflate(R.layout.fragment_activity, container, false)
-        val tabLayout = myFragment.findViewById<TabLayout>(R.id.tab_layout)
-        val viewPager2 = myFragment.findViewById<ViewPager2>(R.id.pager)
-        val adapter = ViewPagerAdapter(childFragmentManager, lifecycle)
-        viewPager2.adapter = adapter
-        TabLayoutMediator(tabLayout, viewPager2) {tab, position->
-            when(position) {
-                0->{
-                    tab.text = "Моя"
-                }
-                1->{
-                    tab.text = "Пользователей"
-                }
-            }
-        }.attach()
-        return myFragment
-    }
-}
